ML_Backend/pair_people/requirements.txt
# Core ML/RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.11
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
torch>=2.0.0
chromadb>=0.4.18
numpy>=1.24.0
networkx>=3.0
python-louvain>=0.16  # For community detection

# API dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# For optional serverless deployment
mangum>=0.17.0  # AWS Lambda handler 